+++ .mypy.ini
@@ -15,7 +15,7 @@ strict_equality = True
 extra_checks = True
 
 
-exclude = src/ert/resources
+exclude = (?x)(src/ert/resources | src/_ert/forward_model_runner)
 
 [mypy-scipy.*]
 ignore_missing_imports = True
