+++ src/_ert/forward_model_runner/reporting/base.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-from _ert_forward_model_runner.reporting.message import Message
+from _ert.forward_model_runner.reporting.message import Message
 
 
 class Reporter(ABC):
