+++ src/_ert/forward_model_runner/reporting/event.py
@@ -16,14 +16,13 @@ from _ert.events import (
     ForwardModelStepSuccess,
     event_to_json,
 )
-from _ert.threading import ErtThread
-from _ert_forward_model_runner.client import (
+from _ert.forward_model_runner.client import (
     Client,
     ClientConnectionClosedOK,
     ClientConnectionError,
 )
-from _ert_forward_model_runner.reporting.base import Reporter
-from _ert_forward_model_runner.reporting.message import (
+from _ert.forward_model_runner.reporting.base import Reporter
+from _ert.forward_model_runner.reporting.message import (
     _JOB_EXIT_FAILED_STRING,
     Checksum,
     Exited,
@@ -32,7 +31,8 @@ from _ert_forward_model_runner.reporting.message import (
     Running,
     Start,
 )
-from _ert_forward_model_runner.reporting.statemachine import StateMachine
+from _ert.forward_model_runner.reporting.statemachine import StateMachine
+from _ert.threading import ErtThread
 
 logger = logging.getLogger(__name__)
