+++ src/_ert/forward_model_runner/runner.py
@@ -3,8 +3,8 @@ import json
 import os
 from pathlib import Path
 
-from _ert_forward_model_runner.job import Job
-from _ert_forward_model_runner.reporting.message import Checksum, Finish, Init
+from _ert.forward_model_runner.job import Job
+from _ert.forward_model_runner.reporting.message import Checksum, Finish, Init
 
 
 class ForwardModelRunner:
