+++ tests/unit_tests/ensemble_evaluator/test_ensemble_client.py
@@ -2,8 +2,8 @@ from functools import partial
 
 import pytest
 
+from _ert.forward_model_runner.client import Client, ClientConnectionError
 from _ert.threading import ErtThread
-from _ert_forward_model_runner.client import Client, ClientConnectionError
 
 from .ensemble_evaluator_utils import _mock_ws
