+++ tests/unit_tests/forward_model_runner/test_file_reporter.py
@@ -3,9 +3,9 @@ import os.path
 
 import pytest
 
-from _ert_forward_model_runner.job import Job
-from _ert_forward_model_runner.reporting import File
-from _ert_forward_model_runner.reporting.message import (
+from _ert.forward_model_runner.job import Job
+from _ert.forward_model_runner.reporting import File
+from _ert.forward_model_runner.reporting.message import (
     Exited,
     Finish,
     Init,
